feat: added favicon, some deps, and modiefied chat
