+++ package.json
@@ -42,6 +42,7 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-leaflet": "^3.2.5",
+    "uuid": "^8.3.2",
     "webpack": "^5.68.0",
     "zustand": "^3.6.9"
   },
